:sparkles: Refactor vacancy tests: add creation test and improve deletion scenarios
